Fix popup
