+++ src/index.html
@@ -4,7 +4,9 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.3.0/css/all.min.css" integrity="sha512-SzlrxWUlpfuzQ+pcUCosxcglQRNAq/DZjVsC0lE40xsADsfeQoEypE+enwcOiGjk/bSuGGKHEyjSoQ1zVisanQ==" crossorigin="anonymous" referrerpolicy="no-referrer" />
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.3.0/css/all.min.css"
+    integrity="sha512-SzlrxWUlpfuzQ+pcUCosxcglQRNAq/DZjVsC0lE40xsADsfeQoEypE+enwcOiGjk/bSuGGKHEyjSoQ1zVisanQ=="
+    crossorigin="anonymous" referrerpolicy="no-referrer" />
   <title>My movies</title>
 </head>
 
@@ -23,41 +25,34 @@
     </nav>
   </header>
 
-<main>
-
-</main>
-
-  <section class="popup">
-    <div class="popup-container">
-      <div class="img-div">
+  <main>
+    <section class="popup-container">
+      <div class="popup">
         <h2 class="close-btn">✖️</h2>
         <img class="popup-img" src="" alt="">
-      </div>
-      <div class="popup-content">
-        <h2 class="pop-name"></h2>
-        <div class="description">
-          <p class="label">Type: <span class="type"></span></p>
-          <p class="label">Language: <span class="language"></span></p>
+        <div class="popup-content">
+          <h2 class="popup-name"></h2>
+          <div class="popup-description">
+            <p>Type: <span class="popup-type"></span></p>
+            <p>Language: <span class="popup-language"></span></p>
+          </div>
+          <div class="comment-section">
+            <h3>Comments (2)</h3>
+            <ul class="comments">
+              <li>03/11/2021 Alex: I'd love to watch this Movie.</li>
+              <li>03/12/2021 Mia: I love it.</li>
+            </ul>
+            <h3>Add a comment</h3>
+            <form action="#" class="form">
+              <input type="text" class="name" placeholder="Your name" required>
+              <textarea class="comment" placeholder="Your insights" rows="5" required></textarea>
+              <input type="submit" class="submit" value="Comment">
+            </form>
+          </div>
         </div>
-        <div class="comment-section">
-          <h3>Comments(2)</h3>
-          <p>03/11/2021  Alex: I'd love to watch this Movie.</p>
-          <p>03/12/2021  Mia: I love it.</p>
-        </div>
-        <h3>Add a comment</h3>
-       <div class="comment-form">
-         <form action="#" class="form">
-           <input type="text" class="name" placeholder="Your name" required>
-           <textarea class="comment" placeholder="Your insights" required></textarea>
-           <input type="submit" class="submit" value="Comment">
-         </form>
       </div>
-      </div>
-
-
-    </div>
-
-  </section>
+    </section>
+  </main>
 
   <footer>
     <p>Created by Microverse under CC license</p>
