+++ src/index.js
@@ -14,29 +14,3 @@ const initialize = async () => {
 };
 
 initialize();
-
-const popup = document.querySelector('.popup');
-const main = document.querySelector('main');
-const closeBtn = document.querySelector('.close-btn');
-const popImage = document.querySelector('.popup-img');
-const popTitle = document.querySelector('.pop-name');
-const type = document.querySelector('.type');
-const language = document.querySelector('.language');
-
-const showPopup = (movie) => {
-  popup.classList.toggle('active');
-  main.classList.toggle('hide');
-  popImage.src = movie.image.original;
-  popTitle.innerHTML = movie.name;
-  type.innerHTML = movie.type;
-  language.innerHTML = movie.language;
-};
-
-const closePopup = () => {
-  popup.classList.toggle('active');
-  main.classList.toggle('hide');
-};
-
-closeBtn.addEventListener('click', closePopup);
-
-export default showPopup;
