+++ src/style.css
@@ -56,10 +56,6 @@ main {
   max-width: 1000px;
 }
 
-main.hide {
-  display: none;
-}
-
 .movie-tile {
   width: 30%;
   margin-bottom: 30px;
@@ -109,64 +105,61 @@ main.hide {
   color: white;
 }
 
-.popup {
-  display: none;
-}
+/* Popup */
 
-.popup.active {
-  display: block;
+.popup-container {
+  background-color: rgba(255, 255, 255, 0.5);
+  backdrop-filter: blur(5px);
+  display: none;
+  align-items: center;
   flex-direction: column;
-  background-color: white;
+  width: 100vw;
+  height: 100vh;
+  overflow: auto;
   position: fixed;
   top: 0;
   left: 0;
-  right: 0;
-  height: 100%;
-  width: 100%;
+  padding: 50px;
   z-index: 999;
-  padding: 0.5rem;
-  overflow-y: scroll;
 }
 
-.popup-container {
+.popup-container.active {
   display: flex;
-  align-items: center;
-  flex-direction: column;
 }
 
-.img-div {
-  width: 100%;
+.popup {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 20px;
+  background-color: white;
+  padding: 20px;
+  border: 2px solid;
 }
 
 .close-btn {
-  margin-right: 100px;
-  font-weight: 900;
+  float: right;
+  width: 100%;
   text-align: right;
 }
 
-.close-btn:hover {
-  cursor: pointer;
-}
-
 .popup-img {
-  width: 100%;
-  height: 30vh;
-  object-fit: contain;
+  width: 200px;
+  height: auto;
 }
 
-.pop-name {
-  text-align: center;
-  margin: 20px;
-}
-
-.description {
+.popup-content {
   display: flex;
   flex-direction: column;
   align-items: center;
+  gap: 20px;
+  width: 400px;
 }
 
-.description .label {
-  margin-bottom: 10px;
+.popup-description {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
   font-weight: 600;
 }
 
@@ -175,51 +168,45 @@ main.hide {
   flex-direction: column;
   align-items: center;
   font-weight: 500;
+  gap: 20px;
 }
 
 .comment-section h3 {
   font-size: 20px;
-  margin: 20px;
 }
 
-.comment-section p {
-  margin-bottom: 10px;
-}
-
-.popup-content h3 {
-  font-size: 20px;
-  text-align: center;
-  margin-bottom: 20px;
+.comments {
+  list-style: none;
+  font-size: 16px;
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
 }
 
 form {
   display: flex;
   flex-direction: column;
   align-items: center;
+  gap: 20px;
+  width: 100%;
 }
 
-form .name {
-  padding: 5px;
+form input,
+form textarea {
+  padding: 10px;
   width: 100%;
-  height: 30px;
-  margin-bottom: 25px;
-}
-
-form .comment {
-  padding: 5px;
-  width: 100%;
-  height: 50px;
-  margin-bottom: 25px;
+  font-size: 15px;
 }
 
 form .submit {
-  padding: 8px;
+  padding: 8px 16px;
   border: 3px solid;
   background-color: white;
   box-shadow: 4px 4px black;
   font-size: 18px;
   font-weight: 600;
   cursor: pointer;
+  width: max-content;
 }
 
 form .submit:hover {
